Adds nodeeditor to dependencies
