--- nodeeditor/qss/nodestyle.qss
@@ -1,3 +1,0 @@
-QDMNodeContentWidget { background: transparent; }
-QDMNodeContentWidget QLabel { color: #e0e0e0;  }
-QGraphicsView { selection-background-color: rgb(255, 255, 255); }
